+++ bdk-android/lib/src/androidTest/kotlin/org/bitcoindevkit/PersistenceTest.kt
@@ -50,11 +50,11 @@ class PersistenceTest {
 
     private val descriptor: Descriptor = Descriptor(
         "wpkh(tprv8ZgxMBicQKsPf2qfrEygW6fdYseJDDrVnDv26PH5BHdvSuG6ecCbHqLVof9yZcMoM31z9ur3tTYbSnr1WBqbGX97CbXcmp5H6qeMpyvx35B/84h/1h/0h/0/*)",
-        Network.SIGNET
+        NetworkKind.TEST
     )
     private val changeDescriptor: Descriptor = Descriptor(
         "wpkh(tprv8ZgxMBicQKsPf2qfrEygW6fdYseJDDrVnDv26PH5BHdvSuG6ecCbHqLVof9yZcMoM31z9ur3tTYbSnr1WBqbGX97CbXcmp5H6qeMpyvx35B/84h/1h/0h/1/*)",
-        Network.SIGNET
+        NetworkKind.TEST
     )
 
     // fun `Correctly load wallet from sqlite persistence`() {