+++ bdk-swift/Tests/BitcoinDevKitTests/OfflinePersistenceTests.swift
@@ -3,12 +3,12 @@ import XCTest
 
 final class OfflinePersistenceTests: XCTestCase {
     private let descriptor = try! Descriptor(
-    descriptor: "wpkh(tprv8ZgxMBicQKsPf2qfrEygW6fdYseJDDrVnDv26PH5BHdvSuG6ecCbHqLVof9yZcMoM31z9ur3tTYbSnr1WBqbGX97CbXcmp5H6qeMpyvx35B/84h/1h/0h/0/*)", 
-    network: Network.signet
+        descriptor: "wpkh(tprv8ZgxMBicQKsPf2qfrEygW6fdYseJDDrVnDv26PH5BHdvSuG6ecCbHqLVof9yZcMoM31z9ur3tTYbSnr1WBqbGX97CbXcmp5H6qeMpyvx35B/84h/1h/0h/0/*)", 
+        networkKind: NetworkKind.test
     )
     private let changeDescriptor = try! Descriptor(
         descriptor: "wpkh(tprv8ZgxMBicQKsPf2qfrEygW6fdYseJDDrVnDv26PH5BHdvSuG6ecCbHqLVof9yZcMoM31z9ur3tTYbSnr1WBqbGX97CbXcmp5H6qeMpyvx35B/84h/1h/0h/1/*)", 
-        network: Network.signet
+        networkKind: NetworkKind.test
     )
     var dbFilePath: URL!
 
@@ -44,11 +44,11 @@ final class OfflinePersistenceTests: XCTestCase {
         
         let descriptorPub = try Descriptor(
             descriptor: "wpkh([9122d9e0/84'/1'/0']tpubDCYVtmaSaDzTxcgvoP5AHZNbZKZzrvoNH9KARep88vESc6MxRqAp4LmePc2eeGX6XUxBcdhAmkthWTDqygPz2wLAyHWisD299Lkdrj5egY6/0/*)#zpaanzgu",
-            network: Network.signet
+            networkKind: NetworkKind.test
         )
         let changeDescriptorPub = try Descriptor(
             descriptor: "wpkh([9122d9e0/84'/1'/0']tpubDCYVtmaSaDzTxcgvoP5AHZNbZKZzrvoNH9KARep88vESc6MxRqAp4LmePc2eeGX6XUxBcdhAmkthWTDqygPz2wLAyHWisD299Lkdrj5egY6/1/*)#n4cuwhcy",
-            network: Network.signet
+            networkKind: NetworkKind.test
         )
         
         let wallet = try Wallet.load(